kernel/legacy: add partial implementation of <atomic>
This is the implementation of std::atomic<>, std::atomic_signal_fence()
and std::atomic_thread_fence() that works with GCC2 and is sufficiently
similar to the proper C++11 implementation that in most cases can be
uses interchangeably.

The purpose of this patch is to remove any excuses for using old
atomic_*() functions in the new code, std::atomic<> is a more convenient
interface and, unlike old atomic_*() functions, allows specifying ordering
requirements -- an information useful both for the humans and the compilers.

Signed-off-by: Paweł Dziepak <pdziepak@quarnos.org>